requirements.txt
fastapi
uvicorn
jinja2
python-multipart
passlib[bcrypt]
python-docx